refactor: ReservationService 공통으로 사용하도록 수정
- RoomEscapeService -> ReservationService 네이밍 수정
- 콘솔의 어플리케이션에서 파싱 코드 함수로 분리해 Service 및 DTO 공통 사용하도록 수정
- 서비스 패키지 이동으로 인해 컴포넌트 스캔 위한 웹 어플리케이션의 어노테이션 추가
