+++ src/main/java/nextstep/web/RoomEscapeApplication.java
@@ -2,7 +2,9 @@ package nextstep.web;
 
 import org.springframework.boot.SpringApplication;
 import org.springframework.boot.autoconfigure.SpringBootApplication;
+import org.springframework.context.annotation.ComponentScan;
 
+@ComponentScan({"nextstep.service","nextstep.web"})
 @SpringBootApplication
 public class RoomEscapeApplication {
     public static void main(String[] args) {
